feat: integracion de vite.js como empaquetador
