+++ public/index.html
@@ -17,6 +17,6 @@
   </div>
 </body>
 
-<script type="text/javascript" src="../src/index.js"></script>
+<script type="module" src="index.js"></script>
 
 </html>
