+++ vite.config.js
@@ -1,0 +1,9 @@
+import path from "path";
+
+export default {
+  root: path.resolve(__dirname, "public"),
+  build: {
+    outDir: path.resolve(__dirname, "dist"),
+    emptyOutDir: true
+  },
+}
